Move old directory branch widget back inside the file dialog module.
